+++ sui_core/src/consensus_adapter.rs
@@ -18,8 +18,8 @@ use tokio::task::JoinHandle;
 use tracing::debug;
 
 #[cfg(test)]
-#[path = "unit_tests/consensus_adapter_tests.rs"]
-pub mod consensus_adapter_tests;
+#[path = "unit_tests/consensus_tests.rs"]
+pub mod consensus_tests;
 
 /// A serialized consensus transaction.
 type SerializedConsensusTransaction = Vec<u8>;
